move the event.Bus to eventbus.Bus, deprecate old types
