+++ eventbus/bus.go
@@ -1,4 +1,4 @@
-package event
+package eventbus
 
 import (
 	"io"
